Add day 2 solution

+++ 2025/joyful-velvet-mistletoe/src/main/java/aoc/twentyfive/puzzles/PuzzleSolver.java
@@ -1,14 +1,12 @@
 package aoc.twentyfive.puzzles;
 
-import aoc.twentyfive.comon.InputUtil;
-
 import static java.lang.IO.println;
 
 abstract class PuzzleSolver<Input> {
 
     public abstract Input parseInput();
-    public abstract int part1(Input input);
-    public abstract int part2(Input input);
+    public abstract long part1(Input input);
+    public abstract long part2(Input input);
 
     void main() {
         var input = parseInput();
